+++ src/Pool/WeightedPool.php
@@ -33,7 +33,7 @@ final readonly class WeightedPool implements PoolInterface
     /**
      * @template TItem
      * @param iterable<TItem>                    $items
-     * @param \Closure(TItem): int               $weightFn
+     * @param \Closure(TItem): int               $weightExtractor
      * @param ItemFilterInterface<TItem>          $filter
      * @param class-string<SelectorInterface>     $selectorClass
      * @param RandomizerInterface                 $randomizer
@@ -41,7 +41,7 @@ final readonly class WeightedPool implements PoolInterface
      */
     public static function of(
         iterable $items,
-        \Closure $weightFn,
+        \Closure $weightExtractor,
         ItemFilterInterface $filter = new PositiveValueFilter(),
         string $selectorClass = PrefixSumSelector::class,
         RandomizerInterface $randomizer = new SecureRandomizer(),
@@ -51,7 +51,7 @@ final readonly class WeightedPool implements PoolInterface
         /** @var list<int> $filteredWeights */
         $filteredWeights = [];
         foreach ($items as $item) {
-            $weight = $weightFn($item);
+            $weight = $weightExtractor($item);
             if ($filter->accepts($item, $weight)) {
                 $filteredItems[]   = $item;
                 $filteredWeights[] = $weight;